Modernize conxian-nexus to v0.4.11 production baseline
- Upgrade core dependencies: axum v0.8, sqlx v0.9, redis v0.27, nostr-sdk v0.43.0.
- Centralize all environment variable access in `src/config.rs`.
- Harden security with parameterized SQL and secret redaction in logs.
- Refactor Nostr and Kwil modules for protocol-first alignment.
- Update all integration tests to match modernized API signatures.
- Globally align versioning to 0.4.11.

+++ src/api/settlement.rs
@@ -164,8 +164,8 @@ pub async fn settlement_trigger_handler(
                 source: payload.source.clone(),
                 payload: payload.payload.clone(),
                 status: "active".to_string(),
-                init_height: current_height,
-                unlock_height: unlock_height as i64,
+                init_height: current_height as u64,
+                unlock_height: unlock_height as u64,
             })
             .await
             .map_err(|e| tracing::warn!("Kwil settlement proposal persistence failed: {}", e))
